--- src/Components/BrainTree.css
@@ -1,21 +1,0 @@
-#card-number {
-  border: 1px solid #333;
-  -webkit-transition: border-color 160ms;
-  transition: border-color 160ms;
-}
-
-#card-number.braintree-hosted-fields-focused {
-  border-color: #777;
-}
-
-#card-number.braintree-hosted-fields-invalid {
-  border-color: tomato;
-}
-
-#card-number.braintree-hosted-fields-valid {
-  border-color: limegreen;
-}
-
-.background {
-  background-color: aqua;
-}
