+++ src/Views/Paypal/Paypal.jsx
@@ -10,7 +10,7 @@ import PropTypes from 'prop-types';
 import * as Styles from './Paypal.style';
 import * as actions from '../../store/actions';
 // import PaypalButton from '../../Components/paypal';
-import BrainTree from '../../Components/BrainTree';
+import BrainTree from '../../Components/Braintree';
 
 function Paypal({ payment, error, paymentAction }) {
   const [successState, setSuccessState] = useState(false);
